fix(release): own Windows smoke homes explicitly
