+++ desktop/ambient-orb/scripts/build-contract.mjs
@@ -32,6 +32,7 @@ const BUILD_JAVASCRIPT_FILES = Object.freeze([
   'scripts/sign-mac-with-native-manifest.cjs',
   'scripts/after-sign.cjs',
   'scripts/installed-candidate-smoke.mjs',
+  'scripts/windows-smoke-home.mjs',
   'scripts/collect-release-artifacts.mjs',
   'scripts/generate-pending-release-ledger.mjs',
   'scripts/generate-release-candidate-report.mjs',
