+++ desktop/ambient-orb/scripts/installed-candidate-smoke.mjs
@@ -3,11 +3,19 @@ import {createHash} from 'node:crypto'
 import {spawn, spawnSync} from 'node:child_process'
 import {createServer as createHttpsServer} from 'node:https'
 import {copyFile, lstat, mkdir, mkdtemp, readFile, realpath, rm, stat} from 'node:fs/promises'
-import {homedir, tmpdir} from 'node:os'
 import {isAbsolute, join, posix, resolve, win32} from 'node:path'
 import {fileURLToPath} from 'node:url'
 
 import {WebSocket, WebSocketServer} from 'ws'
+import {
+  candidateScratchParent,
+  prepareWindowsSmokeHomeOwnership,
+} from './windows-smoke-home.mjs'
+
+export {
+  candidateScratchParent,
+  prepareWindowsSmokeHomeOwnership,
+} from './windows-smoke-home.mjs'
 
 export const RELEASE_SMOKE_MODE = 'installed-candidate-v1'
 export const CAMERA_CAPABILITY_PENDING = 'camera-file-integration: chromium_codec_unavailable'
@@ -21,55 +29,6 @@ export const SCRATCH_REMOVAL_OPTIONS = Object.freeze({
   maxRetries: 100,
   retryDelay: 50,
 })
-
-export function candidateScratchParent({
-  platform = process.platform,
-  home = homedir(),
-  temporary = tmpdir(),
-} = {}) {
-  const parent = platform === 'win32' ? home : temporary
-  const pathApi = platform === 'win32' ? win32 : posix
-  if (typeof parent !== 'string' || !pathApi.isAbsolute(parent)) {
-    throw new Error('installed_candidate_invalid')
-  }
-  return pathApi.resolve(parent)
-}
-
-export function prepareWindowsSmokeHomeOwnership({
-  platform = process.platform,
-  home,
-  environment,
-  run = spawnSync,
-}) {
-  if (platform !== 'win32') return
-  const systemRoot = environment?.SystemRoot ?? environment?.WINDIR
-  if (typeof home !== 'string' || !win32.isAbsolute(home)
-    || typeof systemRoot !== 'string' || !win32.isAbsolute(systemRoot)) {
-    throw new Error('installed_candidate_invalid')
-  }
-  const options = {
-    env: environment,
-    encoding: 'utf8',
-    windowsHide: true,
-    timeout: 10_000,
-    maxBuffer: OUTPUT_LIMIT,
-    stdio: ['ignore', 'pipe', 'pipe'],
-  }
-  const identity = run(win32.join(systemRoot, 'System32', 'whoami.exe'), [
-    '/user', '/fo', 'csv', '/nh',
-  ], options)
-  const match = typeof identity?.stdout === 'string'
-    ? /^"[^"\r\n]{1,256}","(S-1-(?:[0-9]+-){1,15}[0-9]+)"\r?\n?$/u.exec(identity.stdout)
-    : null
-  if (identity?.status !== 0 || identity?.error !== undefined || identity?.signal !== null
-    || match === null) throw new Error('installed_candidate_invalid')
-  const ownership = run(win32.join(systemRoot, 'System32', 'icacls.exe'), [
-    home, '/setowner', `*${match[1]}`, '/Q',
-  ], options)
-  if (ownership?.status !== 0 || ownership?.error !== undefined || ownership?.signal !== null) {
-    throw new Error('installed_candidate_invalid')
-  }
-}
 
 const DEFAULT_SIGNER_WORKFLOW = 'deepnovacore/NovaAudioAgent/.github/workflows/release-candidate.yml'
 const SIGNER_WORKFLOWS = new Set([
