+++ desktop/ambient-orb/scripts/run-packaged-codex-smoke.mjs
@@ -15,7 +15,8 @@ const DIRECTORY_FAILURE_STAGES = ['open', 'create', 'protect'].flatMap(operation
 ))
 const FAILURE_STAGES = new Set([
   'load_runtime', 'settings_project', 'host_project', 'host_project_transport',
-  'host_project_native', 'host_project_directories', 'host_project_config', 'resource_project',
+  'host_project_native', 'host_project_home', 'host_project_directories',
+  'host_project_config', 'resource_project',
   'resource_project_codex_host_unavailable',
   'resource_project_codex_project_host_unsupported',
   'resource_project_codex_project_state_invalid',
