+++ desktop/ambient-orb/test/release-smoke-certificate.test.mjs
@@ -81,5 +81,6 @@ test('release smoke kit generates its TLS identity instead of copying ignored fi
     'utf8',
   )
   assert.match(source, /generateReleaseSmokeCertificate\(\{/u)
+  assert.match(source, /scripts\/windows-smoke-home\.mjs/u)
   assert.doesNotMatch(source, /resolve\(packageRoot, 'scripts\/release-smoke-(?:cert|key)\.pem'\)/u)
 })
